Split agents module into 3 modules
